Move the files about again to match a different new style
